+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Add default_url_options plugin for specifiying default URL options for uploaded files (janko-m)
+
 * Add module_include plugin for easily extending core classes for given uploader (janko-m)
 
 * Add support for Sequel's Postgres JSON column support (janko-m)